Added Sprite class.
The Sprite class will be used to create different Sprites in the game.

Expanded the capabilites of the Parent Sprite class to allow for more
dynamic drawings using Polygons, the abilities to use Images in Sprites.
This is to allow for more detailed graphics to be drawn to the screen, a
needed aspect of many games.